Bootstrap repository with Spring Boot (#8)
* chore: bootstrap repository with Spring Boot

Add files and necessary libraries necessary and essential for the rest
of the project.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* chore: add further git dotfiles

Add .DS_Store and "**/build" entries to .gitignore and also create a
.gitattributes file.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* chore: add plugins section on libs.versions.toml

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* chore: add spotless to build.gradle

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* chore: add Spring Boot to build.gradle

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* style: add .editorconfig

We need a standardized way to do spacing in our code. Add .editorconfig
to fix that.

It's universal for most IDEs so it should be fine for most developers
working on this project.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

---------

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>