docker-compose.yml: implement auxiliary services (#15)
* docker-compose.yml: implement auxiliary services

For our self-hosted chat application, we are going to need working and
reliable instances of:

  - Apache Kafka

  - Apache Cassandra

  - PostgreSQL

Create a docker-compose.yml file as well as a Dockerfile to package our
code into a JAR and from that image bundle it with all the
aforementioned services with Docker Compose.

There is a high chance that some of the environment variables are not
what they are supposed to be, but further clarifications need to be made
about where to get them from.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* chore: use .env and update docker-compose.yml

Separate all environment variables into a .env file for users to set up
everything from there and remove some unnecessary stuff from
docker-compose.yml.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* docker-compose.yml: remove Cassandra network

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

---------

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>