Reinforce indentation on non-Java files (#16)
* style: reinforce indentation on non-Java files

Great, we already have the infrastructure for standardizing indentation
and styling for our Java files and code in general, but now we need to
have the same technology working for our configuration files.

In build.gradle, reinforce the styling of .gradle files which is mostly
what we work with so far.  Of course,  more file types are to be added
for ones that we start using in the project, we just can't use a
wildcard for all filetypes and hope that it is going to work.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

* fix: oops, foreshadowing, not supposed to be here

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>

---------

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>