authorization-microservice(oauth): lengthen token expiry
It's currently so annoying to debug all of the endpoints unrelated to
this mircoservice because we have to use the refresh token as part of
the OAuth 2.0 flow and it really slows down testing.

Set the token TTL to the duration of 1 day to reduce the frequency of
the process of refreshing the token.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>